+++ pro-mini/include/adc.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-int adc_init(void);
+uint8_t adc_init(void);
 uint16_t adc_read(uint8_t channel);
 
 #endif
